moved DemoError and RemoteParticipantAvatarHelper in the same bridging file for simplicity.
