--- AzureCommunicationUIDemoApp/ios/DemoError.swift
@@ -1,17 +1,0 @@
-//
-//  Copyright (c) Microsoft Corporation. All rights reserved.
-//  Licensed under the MIT License.
-//
-
-import AzureCommunicationUICalling
-
-enum DemoError: Error {
-    case invalidToken
-
-    func getErrorCode() -> String {
-        switch self {
-        case .invalidToken:
-            return CallCompositeErrorCode.tokenExpired
-        }
-    }
-}
